+++ R/get_contrasts_character.R
@@ -11,7 +11,7 @@ get_contrast_data_character <- function(model,
 
     levs_idx <- data.table::data.table(lo = levs[1], hi = levs[2:length(levs)])
     levs_idx$label <- sprintf("%s - %s", levs_idx$hi, levs_idx$lo)
-    levs_idx <- setNames(levs_idx, paste0("marginaleffects_contrast_", colnames(levs_idx)))
+    levs_idx <- stats::setNames(levs_idx, paste0("marginaleffects_contrast_", colnames(levs_idx)))
 
 
     # https://github.com/Rdatatable/data.table/issues/1717#issuecomment-545758165