+++ R/get_jacobian.R
@@ -5,25 +5,17 @@ get_jacobian <- function(func, x) {
     # global options are used to switch between homemade and numDeriv
     marginaleffects_numDeriv <- getOption("marginaleffects_numDeriv", default = NULL)
     flag <- is.null(marginaleffects_numDeriv)
+    eps <- getOption("marginaleffects_deriv_eps", default = 0.0001)
 
     # dependency-free
     if (isTRUE(flag)) {
-        eps <- 1e-4
         baseline <- func(x)
-        # df <- matrix(NA, length(baseline), length(x))
-        # for (i in seq_along(x)) {
-        #     dx <- x
-        #     dx[i] <- dx[i] + eps
-        #     df[, i] <- (func(dx) - baseline) / eps
-        # }
-        df <- data.table(matrix(NA_real_, length(baseline), length(x)))
+        df <- matrix(NA_real_, length(baseline), length(x))
         for (i in seq_along(x)) {
             dx <- x
             dx[i] <- dx[i] + eps
-            df[, (i) := (func(dx) - baseline) / eps]
+            df[, i] <- (func(dx) - baseline) / eps
         }
-        df <- as.matrix(df)
-        colnames(df) <- NULL
 
     # numDeriv package (more flexible)
     } else {