+++ R/sanity_dots.R
@@ -5,18 +5,11 @@ sanity_dots <- function(model, ...) {
     # mixed effects
     valid[["merMod"]] <- valid[["lmerMod"]] <- valid[["glmerMod"]] <- valid[["lmerModLmerTest"]] <-
         c("include_random", "re.form", "allow.new.levels", "random.only")
-
-    # bayesian
     valid[["brmsfit"]] <- c("ndraws", "re_formula", "allow_new_levels", "sample_new_levels")
-
-    # sampleSelection
-    valid[["selection"]] <- c("part")
-
-    # glmmTMB
-    valid[["glmmTMB"]] <- c("re.form", "allow.new.levels", "zitype")
-
-    # misc
-    valid[["bam"]] <- c("exclude")
+    valid[["selection"]] <- c("part") # sampleSelection
+    valid[["glmmTMB"]] <- c("re.form", "allow.new.levels", "zitype") # glmmTMB
+    valid[["bam"]] <- c("exclude") # mgcv
+    valid[["rlmerMod"]] <- c("re.form", "allow.new.levels")
 
     white_list <- c("conf.int")