+++ R/standard_errors_delta.R
@@ -24,23 +24,9 @@ standard_errors_delta_contrasts <- function(model,
                   type = type,
                   contrast_factor = contrast_factor,
                   contrast_numeric = contrast_numeric,
-                  vcov = FALSE,
                   ...)$comparison
 }
 
-
-standard_errors_delta_marginaleffects <- function(model,
-                                                  variable,
-                                                  newdata,
-                                                  type,
-                                                  ...) {
-    get_dydx(model = model,
-             variable = variable,
-             newdata = newdata,
-             type = type,
-             vcov = FALSE,
-             ...)$dydx
-}
 
 
 #' Compute standard errors using the delta method