+++ tests/testthat/test-pkg-rms.R
@@ -15,11 +15,10 @@ test_that("lmr: marginaleffects vs emtrends", {
 })
 
 test_that("predictions: rms: no validity", {
-    skip("unsupported data argument in get_predicted.lrm")
     model <- rms::lrm(am ~ mpg, mtcars)
     pred1 <- predictions(model, type = "lp")
     pred2 <- predictions(model, type = "lp", newdata = head(mtcars))
-    expect_predictions(pred1, n_row = 1)
+    expect_predictions(pred1, n_row = 32)
     expect_predictions(pred2, n_row = 6)
 })